uptime: split platform-specific code into unix and windows submodules
