+++ src/uu/uptime/src/uptime.rs
@@ -3,15 +3,13 @@
 // For the full copyright and license information, please view the LICENSE
 // file that was distributed with this source code.
 
-// spell-checker:ignore getloadavg behaviour loadavg uptime upsecs updays upmins uphours boottime nusers utmpxname gettime clockid couldnt
+// spell-checker:ignore behaviour loadavg nusers
+
+mod platform;
 
 use clap::{Arg, ArgAction, Command};
-#[cfg(unix)]
-use clap::{ValueHint, builder::ValueParser};
 use jiff::tz::TimeZone;
 use jiff::{Timestamp, ToSpan};
-#[cfg(unix)]
-use std::ffi::OsString;
 use std::io::{self, Write, stdout};
 use thiserror::Error;
 use uucore::error::{UError, UResult};
@@ -20,12 +18,8 @@ use uucore::libc::time_t;
 use uucore::translate;
 use uucore::uptime::{
     OutputFormat, format_nusers, get_formatted_loadavg, get_formatted_nusers, get_formatted_time,
-    get_formatted_uptime, get_uptime,
+    get_formatted_uptime,
 };
-
-#[cfg(unix)]
-#[cfg(not(target_os = "openbsd"))]
-use uucore::utmpx::{BOOT_TIME, USER_PROCESS, Utmpx};
 
 pub mod options {
     pub static SINCE: &str = "since";
@@ -60,9 +54,8 @@ pub fn uumain(args: impl uucore::Args) -> UResult<()> {
     if matches.get_flag(options::PRETTY) {
         return pretty_print_uptime();
     }
-    #[cfg(unix)]
-    if let Some(path) = matches.get_one::<OsString>(options::PATH) {
-        return uptime_with_file(path);
+    if let Some(result) = platform::maybe_uptime_from_file(&matches) {
+        return result;
     }
     default_uptime()
 }
@@ -93,122 +86,11 @@ pub fn uu_app() -> Command {
                 .help(translate!("uptime-help-pretty"))
                 .action(ArgAction::SetTrue),
         );
-    // The utmp file operand only makes sense where utmp files exist.
-    #[cfg(unix)]
-    let cmd = cmd.arg(
-        Arg::new(options::PATH)
-            .help(translate!("uptime-help-path"))
-            .action(ArgAction::Set)
-            .num_args(0..=1)
-            .value_parser(ValueParser::os_string())
-            .value_hint(ValueHint::AnyPath),
-    );
-    cmd
-}
-
-#[cfg(unix)]
-fn uptime_with_file(file_path: &OsString) -> UResult<()> {
-    use std::fs;
-    use std::os::unix::fs::FileTypeExt;
-    use uucore::error::set_exit_code;
-    use uucore::show_error;
-
-    // Uptime will print loadavg and time to stderr unless we encounter an extra operand.
-    let mut non_fatal_error = false;
-
-    // process_utmpx_from_file() doesn't detect or report failures, we check if the path is valid
-    // before proceeding with more operations.
-    let md_res = fs::metadata(file_path);
-    if let Ok(md) = md_res {
-        if md.is_dir() {
-            show_error!("{}", UptimeError::TargetIsDir);
-            non_fatal_error = true;
-            set_exit_code(1);
-        }
-        if md.file_type().is_fifo() {
-            show_error!("{}", UptimeError::TargetIsFifo);
-            non_fatal_error = true;
-            set_exit_code(1);
-        }
-    } else if let Err(e) = md_res {
-        non_fatal_error = true;
-        set_exit_code(1);
-        show_error!("{}", UptimeError::IoErr(e));
-    }
-    // utmpxname() returns an -1 , when filename doesn't end with 'x' or its too long.
-    // Reference: `<https://developer.apple.com/library/archive/documentation/System/Conceptual/ManPages_iPhoneOS/man3/utmpxname.3.html>`
-
-    #[cfg(target_os = "macos")]
-    {
-        use std::os::unix::ffi::OsStrExt;
-        let bytes = file_path.as_os_str().as_bytes();
-
-        if bytes[bytes.len() - 1] != b'x' {
-            show_error!("{}", translate!("uptime-error-couldnt-get-boot-time"));
-            print_time()?;
-            write!(stdout(), "{}", translate!("uptime-output-unknown-uptime"))?;
-            print_nusers(Some(0))?;
-            print_loadavg()?;
-            set_exit_code(1);
-            return Ok(());
-        }
-    }
-
-    if non_fatal_error {
-        print_time()?;
-        write!(stdout(), "{}", translate!("uptime-output-unknown-uptime"))?;
-        print_nusers(Some(0))?;
-        print_loadavg()?;
-        return Ok(());
-    }
-
-    print_time()?;
-    let user_count;
-
-    #[cfg(not(target_os = "openbsd"))]
-    {
-        let (boot_time, count) = process_utmpx(Some(file_path));
-        if let Some(time) = boot_time {
-            print_uptime(Some(time))?;
-        } else {
-            show_error!("{}", translate!("uptime-error-couldnt-get-boot-time"));
-            set_exit_code(1);
-
-            write!(stdout(), "{}", translate!("uptime-output-unknown-uptime"))?;
-        }
-        user_count = count;
-    }
-
-    #[cfg(target_os = "openbsd")]
-    {
-        let upsecs = get_uptime(None)?;
-        if upsecs >= 0 {
-            print_uptime(Some(upsecs))?;
-        } else {
-            show_error!("{}", translate!("uptime-error-couldnt-get-boot-time"));
-            set_exit_code(1);
-
-            write!(stdout(), "{}", translate!("uptime-output-unknown-uptime"))?;
-        }
-        user_count =
-            uucore::uptime::get_nusers(file_path.to_str().expect("invalid utmp path file"));
-    }
-
-    print_nusers(Some(user_count))?;
-    print_loadavg()?;
-
-    Ok(())
+    platform::add_platform_args(cmd)
 }
 
 fn uptime_since() -> UResult<()> {
-    #[cfg(unix)]
-    #[cfg(not(target_os = "openbsd"))]
-    let uptime = {
-        let (boot_time, _) = process_utmpx(None);
-        get_uptime(boot_time)?
-    };
-    #[cfg(any(windows, target_os = "openbsd"))]
-    let uptime = get_uptime(None)?;
+    let uptime = platform::system_uptime_seconds()?;
 
     let since_date = (Timestamp::now() - uptime.seconds()).to_zoned(TimeZone::system());
     writeln!(stdout(), "{}", since_date.strftime("%Y-%m-%d %H:%M:%S"))?;
@@ -235,32 +117,6 @@ fn print_loadavg() -> UResult<()> {
     }
     writeln!(stdout())?;
     Ok(())
-}
-
-#[cfg(unix)]
-#[cfg(not(target_os = "openbsd"))]
-fn process_utmpx(file: Option<&OsString>) -> (Option<time_t>, usize) {
-    let mut nusers = 0;
-    let mut boot_time = None;
-
-    let records = match file {
-        Some(f) => Utmpx::iter_all_records_from(f),
-        None => Utmpx::iter_all_records(),
-    };
-
-    for line in records {
-        match line.record_type() {
-            x if x == USER_PROCESS => nusers += 1,
-            x if x == BOOT_TIME => {
-                let dt = line.login_time();
-                if dt.unix_timestamp() > 0 {
-                    boot_time = Some(dt.unix_timestamp() as time_t);
-                }
-            }
-            _ => (),
-        }
-    }
-    (boot_time, nusers)
 }
 
 fn print_nusers(nusers: Option<usize>) -> UResult<()> {
